[ADD] :sparkles: long int in entity uuid

+++ rtype_server/src/RtypeServer/RtypeServerCallbacks.cpp
@@ -46,8 +46,8 @@ void RtypeServer::initCallbacks() {
         auto player =
             Player(uuid, elapsed_time, _game_instance->addAndGetPlayer(uuid, PLAYER_START_POSITION),
                    _game_instance->getRegistry());
-        _players[sender] = player;
-        int entity_uuid  = uuid_generator.generate_long();
+        _players[sender]     = player;
+        long int entity_uuid = uuid_generator.generate_long();
         // _server_entities[entity_uuid] =
         // _game_instance->addAndGetEntity({500.f, 200.f});
         std::cout << "Added entity with uuid: " << entity_uuid << std::endl;
